refactor(customer): extrai Customer360 god-component (1248→715)
4º refactor da série de god-components (§10) — 1º fora do domínio Reposição.
Extrai helpers, hooks de dados e subcomponentes de Customer360.tsx para
src/components/customer360/. Extração pura, behavior-preserving:

- format.ts      — formatBRL, formatPctMaybe, splitEmails, formatCustomerType,
                   orderStatusTone, formatDateOrDash, formatRelative, initials,
                   healthTone, churnTone, formatDocument, formatPhone, formatCep
- hooks.ts       — useCustomerCore/Address/Metrics/Score/PreferredItems/Orders/Interactions
- components.tsx — KpiCard, DataRow, ContactRow, ScoreRow

A página vira orquestrador (715 linhas). Import block enxuto. Revisado via codex
review (extração fiel, strict passa, sem regressão). tsc + strict + 392/392 + lint 0.
