+++ src/pages/Customer360.tsx
@@ -1,8 +1,6 @@
 import { useMemo } from 'react';
 import { Link, useNavigate, useParams } from 'react-router-dom';
-import { useQuery } from '@tanstack/react-query';
-import { format, formatDistanceToNow, parseISO, subMonths } from 'date-fns';
-import { ptBR } from 'date-fns/locale';
+import { parseISO, subMonths } from 'date-fns';
 import {
   ArrowLeft,
   Building2,
@@ -22,11 +20,7 @@ import {
   User,
   Inbox,
   Users,
-  Star,
-  Award,
-  Cake,
 } from 'lucide-react';
-import { supabase } from '@/integrations/supabase/client';
 import { useAuth } from '@/contexts/AuthContext';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Badge } from '@/components/ui/badge';
@@ -37,370 +31,31 @@ import { EmptyState } from '@/components/EmptyState';
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip';
 import { cn } from '@/lib/utils';
 import { useCustomerContacts } from '@/hooks/useCustomerContacts';
-import { CARGO_LABEL } from '@/lib/customer-contact/types';
-
-/**
- * Customer 360 — dashboard executivo do cliente.
- *
- * Mostra de relance o que importa pra decidir o próximo passo comercial:
- *  - Identidade + saúde (health/margem/churn)
- *  - Pulso financeiro (12m, ticket, dias desde última compra)
- *  - Itens preferidos (o que ele compra de verdade)
- *  - Último contato (calls + WhatsApp) — abre o caminho pra agir
- *
- * Schema usado:
- *  profiles, addresses, omie_clientes, customer_metrics_mv, farmer_client_scores,
- *  customer_preferred_items, sales_orders, farmer_calls.
- *
- * Filosofia de UX: densidade alta (B2B operacional), KPI em mono pra leitura tabular,
- * cards quase-neutros, status só onde a cor agrega informação (saúde, churn).
- */
-
-const BRL = new Intl.NumberFormat('pt-BR', { style: 'currency', currency: 'BRL' });
-
-function formatBRL(v: number | null | undefined): string {
-  if (v === null || v === undefined || Number.isNaN(v)) return '—';
-  return BRL.format(v);
-}
-
-function formatPctMaybe(v: number | null | undefined): string {
-  if (v === null || v === undefined || Number.isNaN(v)) return '—';
-  // Schema: gross_margin_pct vem como número (ex: 0.32 ou 32 — normalizamos)
-  const pct = v > 1 ? v : v * 100;
-  // Zero sem decimal ("0%" em vez de "0.0%"); outros valores com 1 decimal só se houver fração.
-  if (pct === 0) return '0%';
-  const rounded = Math.round(pct);
-  return Math.abs(pct - rounded) < 0.05 ? `${rounded}%` : `${pct.toFixed(1)}%`;
-}
-
-/**
- * Profiles tem campo `email` text livre. Sincronização do Omie costuma concatenar
- * múltiplos emails com vírgula (ex: "compras@x.com,financeiro@x.com"). Quebra em
- * lista pra renderizar separadamente.
- */
-function splitEmails(raw: string | null | undefined): string[] {
-  if (!raw) return [];
-  return raw
-    .split(/[,;]/)
-    .map((e) => e.trim())
-    .filter((e) => e.length > 0 && e.includes('@'));
-}
-
-/**
- * `profiles.customer_type` vem como enum-ish text do app/Omie. Mapeia pros rótulos
- * que o usuário humano entende. Fallback: capitaliza o valor cru.
- */
-function formatCustomerType(t: string | null | undefined): string {
-  if (!t) return '—';
-  const map: Record<string, string> = {
-    domestic: 'Doméstico',
-    residential: 'Residencial',
-    residencial: 'Residencial',
-    commercial: 'Comercial',
-    comercial: 'Comercial',
-    industrial: 'Industrial',
-    revenda: 'Revenda',
-    reseller: 'Revenda',
-    professional: 'Profissional',
-    profissional: 'Profissional',
-  };
-  return map[t.toLowerCase()] ?? (t.charAt(0).toUpperCase() + t.slice(1));
-}
-
-/** Tom semântico pro badge de status do pedido. Mapeia status string pra cor sutil. */
-function orderStatusTone(status: string): { className: string; label: string } {
-  const s = (status || '').toLowerCase();
-  if (['faturado', 'concluido', 'pago', 'entregue'].includes(s)) {
-    return { className: 'bg-status-success-bg text-status-success-bold border-status-success/20', label: status };
-  }
-  if (['cancelado', 'recusado', 'erro'].includes(s)) {
-    return { className: 'bg-status-error-bg text-status-error-bold border-status-error/20', label: status };
-  }
-  if (['pendente', 'aguardando', 'em_analise'].includes(s)) {
-    return { className: 'bg-status-warning-bg text-status-warning-bold border-status-warning/20', label: status };
-  }
-  return { className: 'text-muted-foreground', label: status };
-}
-
-function formatDateOrDash(d: string | null | undefined): string {
-  if (!d) return '—';
-  try {
-    return format(parseISO(d), "dd 'de' MMM yyyy", { locale: ptBR });
-  } catch {
-    return '—';
-  }
-}
-
-function formatRelative(d: string | null | undefined): string {
-  if (!d) return '—';
-  try {
-    return formatDistanceToNow(parseISO(d), { locale: ptBR, addSuffix: true });
-  } catch {
-    return '—';
-  }
-}
-
-function initials(name: string): string {
-  return name
-    .split(' ')
-    .filter(Boolean)
-    .slice(0, 2)
-    .map((p) => p[0]?.toUpperCase() ?? '')
-    .join('') || '?';
-}
-
-function healthTone(healthClass: string | null): {
-  label: string;
-  className: string;
-  dot: string;
-} {
-  switch (healthClass) {
-    case 'saudavel':
-      return { label: 'Saudável', className: 'text-status-success-bold', dot: 'bg-status-success' };
-    case 'atencao':
-      return { label: 'Atenção', className: 'text-status-warning-bold', dot: 'bg-status-warning' };
-    case 'risco':
-      return { label: 'Em risco', className: 'text-status-error-bold', dot: 'bg-status-error' };
-    case 'critico':
-      return { label: 'Crítico', className: 'text-status-error-bold', dot: 'bg-status-error' };
-    default:
-      return { label: 'Sem score', className: 'text-muted-foreground', dot: 'bg-muted-foreground' };
-  }
-}
-
-function churnTone(risk: number | null): { label: string; className: string } {
-  if (risk === null || risk === undefined) return { label: '—', className: 'text-muted-foreground' };
-  const pct = risk > 1 ? risk : risk * 100;
-  if (pct >= 70) return { label: `${pct.toFixed(0)}% risco churn`, className: 'text-status-error-bold' };
-  if (pct >= 40) return { label: `${pct.toFixed(0)}% risco churn`, className: 'text-status-warning-bold' };
-  return { label: `${pct.toFixed(0)}% risco churn`, className: 'text-status-success-bold' };
-}
-
-/* ─────────────────────────────  Hooks  ───────────────────────────── */
-
-function useCustomerCore(customerId: string | undefined) {
-  return useQuery({
-    queryKey: ['c360-core', customerId],
-    enabled: !!customerId,
-    staleTime: 60_000,
-    queryFn: async () => {
-      const { data, error } = await supabase
-        .from('profiles')
-        .select('user_id, name, email, phone, document, customer_type, cnae, requires_po, created_at, avatar_url, is_approved')
-        .eq('user_id', customerId!)
-        .maybeSingle();
-      if (error) throw error;
-      return data;
-    },
-  });
-}
-
-function useCustomerAddress(customerId: string | undefined) {
-  return useQuery({
-    queryKey: ['c360-address', customerId],
-    enabled: !!customerId,
-    staleTime: 5 * 60_000,
-    queryFn: async () => {
-      const { data } = await supabase
-        .from('addresses')
-        .select('label, street, number, complement, neighborhood, city, state, zip_code, is_default')
-        .eq('user_id', customerId!)
-        .order('is_default', { ascending: false });
-      return data ?? [];
-    },
-  });
-}
-
-function useCustomerMetrics(customerId: string | undefined) {
-  return useQuery({
-    queryKey: ['c360-metrics', customerId],
-    enabled: !!customerId,
-    staleTime: 60_000,
-    queryFn: async () => {
-      const { data } = await supabase
-        .from('customer_metrics_mv')
-        .select('faturamento_90d, faturamento_prev_90d, ticket_medio_90d, pedidos_90d, dias_desde_ultima_compra, intervalo_medio_dias, ultima_compra_data, is_cold_start')
-        .eq('customer_user_id', customerId!)
-        .maybeSingle();
-      return data;
-    },
-  });
-}
-
-function useCustomerScore(customerId: string | undefined, farmerId: string | undefined) {
-  return useQuery({
-    queryKey: ['c360-score', customerId, farmerId],
-    enabled: !!customerId && !!farmerId,
-    staleTime: 60_000,
-    queryFn: async () => {
-      // Tenta o score do farmer atual primeiro; cai pra qualquer score se não existir
-      const { data: own } = await supabase
-        .from('farmer_client_scores')
-        .select('health_score, health_class, churn_risk, expansion_score, priority_score, gross_margin_pct, avg_monthly_spend_180d, days_since_last_purchase, category_count, avg_repurchase_interval, revenue_potential')
-        .eq('customer_user_id', customerId!)
-        .eq('farmer_id', farmerId!)
-        .maybeSingle();
-      if (own) return own;
-      const { data: fallback } = await supabase
-        .from('farmer_client_scores')
-        .select('health_score, health_class, churn_risk, expansion_score, priority_score, gross_margin_pct, avg_monthly_spend_180d, days_since_last_purchase, category_count, avg_repurchase_interval, revenue_potential')
-        .eq('customer_user_id', customerId!)
-        .limit(1)
-        .maybeSingle();
-      return fallback;
-    },
-  });
-}
-
-/** Itens preferidos via Omie (precisa do código Omie do cliente). */
-function useCustomerPreferredItems(customerId: string | undefined) {
-  return useQuery({
-    queryKey: ['c360-preferred', customerId],
-    enabled: !!customerId,
-    staleTime: 5 * 60_000,
-    queryFn: async () => {
-      const { data: link } = await supabase
-        .from('omie_clientes')
-        .select('omie_codigo_cliente')
-        .eq('user_id', customerId!)
-        .maybeSingle();
-      if (!link?.omie_codigo_cliente) return [];
-      const { data } = await supabase
-        .from('customer_preferred_items')
-        .select('product_codigo, product_descricao, familia, order_count, last_ordered_at, account')
-        .eq('omie_codigo_cliente', link.omie_codigo_cliente)
-        .order('last_ordered_at', { ascending: false, nullsFirst: false })
-        .limit(10);
-      return data ?? [];
-    },
-  });
-}
-
-/** Pedidos pra computar faturamento lifetime + 12m (cliente médio: poucos pedidos, OK no client). */
-function useCustomerOrders(customerId: string | undefined) {
-  return useQuery({
-    queryKey: ['c360-orders', customerId],
-    enabled: !!customerId,
-    staleTime: 60_000,
-    queryFn: async () => {
-      const { data } = await supabase
-        .from('sales_orders')
-        .select('id, total, created_at, status, omie_numero_pedido, account')
-        .eq('customer_user_id', customerId!)
-        .order('created_at', { ascending: false })
-        .limit(200);
-      return data ?? [];
-    },
-  });
-}
-
-/** Timeline de contato: calls + mensagens, unificado e ordenado. */
-function useCustomerInteractions(customerId: string | undefined) {
-  return useQuery({
-    queryKey: ['c360-interactions', customerId],
-    enabled: !!customerId,
-    staleTime: 30_000,
-    queryFn: async () => {
-      const [calls, messages] = await Promise.all([
-        supabase
-          .from('farmer_calls')
-          .select('id, started_at, call_type, call_result, is_whatsapp, notes, duration_seconds, revenue_generated, linked_sales_order_id, farmer_id')
-          .eq('customer_user_id', customerId!)
-          .order('started_at', { ascending: false })
-          .limit(15),
-        supabase
-          .from('order_messages')
-          .select('id, created_at, message, is_staff, sender_id, order_id')
-          .in(
-            'order_id',
-            // janela pequena: últimas 10 ordens deste cliente
-            (await supabase
-              .from('orders')
-              .select('id')
-              .eq('user_id', customerId!)
-              .order('created_at', { ascending: false })
-              .limit(10)).data?.map((o) => o.id) ?? [],
-          )
-          .order('created_at', { ascending: false })
-          .limit(15),
-      ]);
-
-      type Item =
-        | { kind: 'call'; at: string; title: string; subtitle: string; tone: string; revenue?: number | null }
-        | { kind: 'message'; at: string; title: string; subtitle: string; tone: string };
-
-      const items: Item[] = [];
-      (calls.data ?? []).forEach((c) => {
-        items.push({
-          kind: 'call',
-          at: c.started_at,
-          title: c.is_whatsapp ? 'WhatsApp enviado' : 'Ligação',
-          subtitle: [c.call_result, c.notes].filter(Boolean).join(' · ').slice(0, 140) || c.call_type,
-          tone: c.call_result === 'contato_sucesso'
-            ? 'text-status-success-bold'
-            : c.call_result === 'sem_resposta'
-              ? 'text-muted-foreground'
-              : 'text-foreground',
-          revenue: c.revenue_generated,
-        });
-      });
-      (messages.data ?? []).forEach((m) => {
-        items.push({
-          kind: 'message',
-          at: m.created_at,
-          title: m.is_staff ? 'Mensagem da equipe' : 'Mensagem do cliente',
-          subtitle: m.message?.slice(0, 140) ?? '',
-          tone: m.is_staff ? 'text-foreground' : 'text-status-info-bold',
-        });
-      });
-      return items.sort((a, b) => (a.at < b.at ? 1 : -1)).slice(0, 20);
-    },
-  });
-}
-
-/* ──────────────────────────  KPI subcomponents  ────────────────────────── */
-
-function KpiCard({
-  label,
-  value,
-  hint,
-  trend,
-  icon: Icon,
-}: {
-  label: string;
-  value: string;
-  hint?: string;
-  trend?: { value: number; label: string };
-  icon?: React.ComponentType<{ className?: string }>;
-}) {
-  return (
-    <Card>
-      <CardContent className="pt-5 pb-4 space-y-1">
-        <div className="flex items-start justify-between gap-2">
-          <span className="text-xs font-medium text-muted-foreground uppercase tracking-wide">
-            {label}
-          </span>
-          {Icon && <Icon className="w-4 h-4 text-muted-foreground" />}
-        </div>
-        <div className="kpi-value">{value}</div>
-        {trend && (
-          <div
-            className={cn(
-              'kpi-delta',
-              trend.value > 0 ? 'text-status-success-bold' : trend.value < 0 ? 'text-status-error-bold' : 'text-muted-foreground',
-            )}
-          >
-            {trend.value > 0 ? '↑' : trend.value < 0 ? '↓' : '·'} {Math.abs(trend.value).toFixed(0)}%
-            <span className="text-muted-foreground font-normal ml-1">{trend.label}</span>
-          </div>
-        )}
-        {hint && !trend && <div className="text-xs text-muted-foreground">{hint}</div>}
-      </CardContent>
-    </Card>
-  );
-}
-
-/* ──────────────────────────────  Page  ────────────────────────────── */
+import {
+  formatBRL,
+  formatPctMaybe,
+  splitEmails,
+  formatCustomerType,
+  orderStatusTone,
+  formatDateOrDash,
+  formatRelative,
+  initials,
+  healthTone,
+  churnTone,
+  formatDocument,
+  formatPhone,
+  formatCep,
+} from '@/components/customer360/format';
+import {
+  useCustomerCore,
+  useCustomerAddress,
+  useCustomerMetrics,
+  useCustomerScore,
+  useCustomerPreferredItems,
+  useCustomerOrders,
+  useCustomerInteractions,
+} from '@/components/customer360/hooks';
+import { KpiCard, DataRow, ContactRow, ScoreRow } from '@/components/customer360/components';
 
 export default function Customer360() {
   const { customerId } = useParams<{ customerId: string }>();
@@ -1057,192 +712,4 @@ export default function Customer360() {
       </div>
     </TooltipProvider>
   );
-}
-
-/* ───────────────────  helpers UI  ─────────────────── */
-
-function DataRow({
-  icon: Icon,
-  label,
-  value,
-  hint,
-  href,
-}: {
-  icon?: React.ComponentType<{ className?: string }>;
-  label: string;
-  value: string | null | undefined;
-  hint?: string;
-  href?: string;
-}) {
-  const display = value ?? '—';
-  return (
-    <div className="flex items-start gap-3">
-      {Icon && <Icon className="w-3.5 h-3.5 text-muted-foreground mt-0.5 shrink-0" />}
-      <div className="flex-1 min-w-0">
-        <div className="text-xs text-muted-foreground">{label}</div>
-        {href && value ? (
-          <a href={href} className="text-sm text-foreground hover:underline truncate block">
-            {display}
-          </a>
-        ) : (
-          <div className="text-sm text-foreground truncate">{display}</div>
-        )}
-        {hint && <div className="text-xs text-muted-foreground mt-0.5">{hint}</div>}
-      </div>
-    </div>
-  );
-}
-
-/**
- * Linha compacta de contato extra (dono / gerente / comprador / etc).
- * Mostra nome + cargo, telefone clicável + WhatsApp, badges (primary, decisão,
- * só WhatsApp, aniversário). Edição completa fica em /admin/customers detail
- * tab — aqui é só leitura pra contexto operacional rápido.
- */
-function ContactRow({
-  contact,
-}: {
-  contact: import('@/lib/customer-contact/types').CustomerContact;
-}) {
-  const displayName = contact.nome ?? formatPhone(contact.phone);
-  const cargoLabel = contact.cargo ? CARGO_LABEL[contact.cargo] : null;
-  const cleanPhone = contact.phone.replace(/\D/g, '');
-  // Aniversário esse mês? Destaque sutil pra lembrar de mandar mensagem.
-  const isBirthdayMonth = (() => {
-    if (!contact.birthday) return false;
-    try {
-      const d = parseISO(contact.birthday);
-      return d.getMonth() === new Date().getMonth();
-    } catch {
-      return false;
-    }
-  })();
-  return (
-    <li className="pt-2 first:pt-0 space-y-1">
-      <div className="flex items-start justify-between gap-2 min-w-0">
-        <div className="min-w-0 flex-1">
-          <div className="flex items-center gap-1.5 flex-wrap">
-            <span className="font-medium text-sm truncate">{displayName}</span>
-            {cargoLabel && (
-              <Badge variant="outline" className="text-[9px] px-1 py-0">
-                {cargoLabel}
-              </Badge>
-            )}
-            {contact.is_primary && (
-              <Tooltip>
-                <TooltipTrigger>
-                  <Star className="w-3 h-3 text-status-warning-bold fill-status-warning-bold" />
-                </TooltipTrigger>
-                <TooltipContent>Contato principal</TooltipContent>
-              </Tooltip>
-            )}
-            {contact.is_decision_maker && (
-              <Tooltip>
-                <TooltipTrigger>
-                  <Award className="w-3 h-3 text-status-info-bold" />
-                </TooltipTrigger>
-                <TooltipContent>Decision maker (quem decide a compra)</TooltipContent>
-              </Tooltip>
-            )}
-            {isBirthdayMonth && (
-              <Tooltip>
-                <TooltipTrigger>
-                  <Cake className="w-3 h-3 text-status-success-bold" />
-                </TooltipTrigger>
-                <TooltipContent>
-                  Aniversário em {format(parseISO(contact.birthday!), "dd 'de' MMM", { locale: ptBR })}
-                </TooltipContent>
-              </Tooltip>
-            )}
-          </div>
-          <div className="flex items-center gap-2 text-xs text-muted-foreground mt-0.5">
-            {contact.whatsapp_only ? (
-              <a
-                href={`https://wa.me/${cleanPhone}`}
-                target="_blank"
-                rel="noopener noreferrer"
-                className="hover:underline hover:text-foreground transition-colors"
-              >
-                {formatPhone(contact.phone)} · só WhatsApp
-              </a>
-            ) : (
-              <a
-                href={`tel:${contact.phone}`}
-                className="hover:underline hover:text-foreground transition-colors"
-              >
-                {formatPhone(contact.phone)}
-              </a>
-            )}
-          </div>
-          {contact.email && (
-            <a
-              href={`mailto:${contact.email}`}
-              className="text-xs text-muted-foreground hover:underline hover:text-foreground transition-colors truncate block mt-0.5"
-            >
-              {contact.email}
-            </a>
-          )}
-        </div>
-        {!contact.whatsapp_only && (
-          <a
-            href={`https://wa.me/${cleanPhone}`}
-            target="_blank"
-            rel="noopener noreferrer"
-            className="text-muted-foreground hover:text-status-success-bold transition-colors shrink-0 mt-0.5"
-            aria-label="Enviar WhatsApp"
-          >
-            <MessageSquare className="w-3.5 h-3.5" />
-          </a>
-        )}
-      </div>
-    </li>
-  );
-}
-
-/**
- * Linha de score (label esquerda, valor à direita em mono). Usada no card "Score
- * comercial" pra densidade tabular — diferente de DataRow que é vertical (form-like).
- */
-function ScoreRow({
-  label,
-  value,
-  hint,
-}: {
-  label: string;
-  value: string;
-  hint?: string;
-}) {
-  return (
-    <div className="flex items-baseline justify-between gap-2">
-      <div className="min-w-0">
-        <div className="text-xs text-muted-foreground truncate">{label}</div>
-        {hint && <div className="text-[10px] text-muted-foreground/70 mt-0.5">{hint}</div>}
-      </div>
-      <div className="font-mono text-sm font-medium text-foreground tabular-nums shrink-0">
-        {value}
-      </div>
-    </div>
-  );
-}
-
-/* ───────────────────  format helpers  ─────────────────── */
-
-function formatDocument(doc: string): string {
-  const d = doc.replace(/\D/g, '');
-  if (d.length === 11) return d.replace(/(\d{3})(\d{3})(\d{3})(\d{2})/, '$1.$2.$3-$4');
-  if (d.length === 14) return d.replace(/(\d{2})(\d{3})(\d{3})(\d{4})(\d{2})/, '$1.$2.$3/$4-$5');
-  return doc;
-}
-
-function formatPhone(p: string): string {
-  const d = p.replace(/\D/g, '');
-  if (d.length === 11) return d.replace(/(\d{2})(\d{5})(\d{4})/, '($1) $2-$3');
-  if (d.length === 10) return d.replace(/(\d{2})(\d{4})(\d{4})/, '($1) $2-$3');
-  return p;
-}
-
-function formatCep(z: string): string {
-  const d = z.replace(/\D/g, '');
-  if (d.length === 8) return d.replace(/(\d{5})(\d{3})/, '$1-$2');
-  return z;
 }
